[ECG|POPT] add epoch parametrized function

+++ 00_graph_bench/include/structures/graphCSR.h
@@ -27,7 +27,7 @@ struct GraphCSR
     struct EdgeList *inverse_sorted_edges_array; // sorted edge array
 #endif
 
-    uint8_t  *offset_matrix;
+    uint32_t *offset_matrix;
 };
 
 void graphCSRFree (struct GraphCSR *graphCSR);
